I/O Graph: Explicitly convert from uint64_t to double, part 2
We don't care about the loss of precision as we're going to graph
it anyway, but MSVC does.

Part two, this doesn't need backporting to 4.4

+++ ui/io_graph_item.c
@@ -157,9 +157,9 @@ double get_io_graph_item(const io_graph_item_t *items_, io_graph_item_unit_t val
     case IOG_ITEM_UNIT_PACKETS:
         return asAOT ? MICROSECS_PER_SEC*item->frames/interval_ : item->frames;
     case IOG_ITEM_UNIT_BYTES:
-        return asAOT ? MICROSECS_PER_SEC*item->bytes/interval_ : item->bytes;
+        return (double)(asAOT ? MICROSECS_PER_SEC*item->bytes/interval_ : item->bytes);
     case IOG_ITEM_UNIT_BITS:
-        return asAOT ? MICROSECS_PER_SEC*item->bytes*8/interval_ : item->bytes*8;
+        return (double)(asAOT ? MICROSECS_PER_SEC*item->bytes*8/interval_ : item->bytes*8);
     case IOG_ITEM_UNIT_CALC_FRAMES:
         return item->frames;
     case IOG_ITEM_UNIT_CALC_FIELDS:
